+++ packages/framework/next-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @igrp/framework-next-ui
 
+## 0.1.0-beta.146
+
+### Patch Changes
+
+- Updated dependencies
+  - @igrp/framework-next-auth@0.1.0-beta.138
+  - @igrp/framework-next-types@0.1.0-beta.140
+
+## 0.1.0-beta.145
+
+### Patch Changes
+
+- Updated dependencies
+  - @igrp/framework-next-auth@0.1.0-beta.137
+  - @igrp/framework-next-types@0.1.0-beta.139
+
 ## 0.1.0-beta.144
 
 ### Patch Changes